Reset when FILTER count is incorrect
Refactor DocumentListener and QueryListener into Watch::Listener to
get the correct count behavior.
