+++ google-cloud-firestore/test/google/cloud/firestore/collection_reference/listen_restart_test.rb
@@ -28,7 +28,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
   end
   before do
     # Remove start method, make no-op
-    class Google::Cloud::Firestore::QueryListener
+    class Google::Cloud::Firestore::Watch::Listener
       alias_method :start_for_realz, :start
       remove_method :start
 
@@ -49,7 +49,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
   after do
     # restore start method, cleanup our methods
-    class Google::Cloud::Firestore::QueryListener
+    class Google::Cloud::Firestore::Watch::Listener
       alias_method :start, :start_for_realz
       remove_method :start_for_realz
       remove_method :sleep
@@ -71,7 +71,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
     # Start listener because we stopped this from happening in this setup
     # Pass in mock to verify incremental backoff is happening
-    listener.start_with_sleep_mock sleep_mock
+    listener.instance_variable_get(:@listener).start_with_sleep_mock sleep_mock
 
     wait_until { query_snapshots.count == 4 }
 
@@ -92,7 +92,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
     # Start listener because we stopped this from happening in this setup
     # Pass in mock to verify incremental backoff is happening
-    listener.start_with_sleep_mock sleep_mock
+    listener.instance_variable_get(:@listener).start_with_sleep_mock sleep_mock
 
     wait_until { query_snapshots.count == 4 }
 
@@ -113,7 +113,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
     # Start listener because we stopped this from happening in this setup
     # Pass in mock to verify incremental backoff is happening
-    listener.start_with_sleep_mock sleep_mock
+    listener.instance_variable_get(:@listener).start_with_sleep_mock sleep_mock
 
     wait_until { query_snapshots.count == 4 }
 
@@ -134,7 +134,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
     # Start listener because we stopped this from happening in this setup
     # Pass in mock to verify incremental backoff is happening
-    listener.start_with_sleep_mock sleep_mock
+    listener.instance_variable_get(:@listener).start_with_sleep_mock sleep_mock
 
     wait_until { query_snapshots.count == 4 }
 
@@ -155,7 +155,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
     # Start listener because we stopped this from happening in this setup
     # Pass in mock to verify incremental backoff is happening
-    listener.start_with_sleep_mock sleep_mock
+    listener.instance_variable_get(:@listener).start_with_sleep_mock sleep_mock
 
     wait_until { query_snapshots.count == 4 }
 
@@ -176,7 +176,7 @@ describe Google::Cloud::Firestore::CollectionReference, :listen, :backoff, :watc
 
     # Start listener because we stopped this from happening in this setup
     # Pass in mock to verify incremental backoff is happening
-    listener.start_with_sleep_mock sleep_mock
+    listener.instance_variable_get(:@listener).start_with_sleep_mock sleep_mock
 
     wait_until { query_snapshots.count == 4 }
